Include dimensionality in serializers interface
Adds 'startArrayDim' and 'endArrayDim' calls on (de)serializers to
better express the shape of rectangular arrays. This logic *could* be
inferred with recursive calls to 'startArray' and 'endArray', but that
proves to be tricky to debug and think about in practice. The extra
method calls are a bit unfortunate, but ultimately help to more easily
capture the array being read or written.

Adds a test demonstrating that we can write from one type (2D block)
and read into another (2D DefaultRectangular). This is a particularly
useful test for the BinaryIO format where Block is written
element-by-element, but DR is written in bulk.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>